Backend/requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
numpy
python-dotenv
requests
chromadb 
groq
python-multipart